+++ lib/neo4j/http.rb
@@ -8,13 +8,17 @@ require "active_support/core_ext/hash/indifferent_access"
 
 require "neo4j/http/auth_token"
 require "neo4j/http/client"
+require "neo4j/http/batch_client"
 require "neo4j/http/configuration"
 require "neo4j/http/cypher_client"
+require "neo4j/http/batch_cypher_client"
 require "neo4j/http/object_wrapper"
 require "neo4j/http/node"
 require "neo4j/http/node_client"
+require "neo4j/http/batch_node_client"
 require "neo4j/http/relationship"
 require "neo4j/http/relationship_client"
+require "neo4j/http/batch_relationship_client"
 require "neo4j/http/results"
 
 require "neo4j/http/errors"
@@ -24,7 +28,7 @@ module Neo4j
     extend self
 
     def config
-      @congiguration ||= Configuration.new
+      @configuration ||= Configuration.new
     end
 
     def configure