+++ lib/neo4j/http/client.rb
@@ -3,7 +3,7 @@
 module Neo4j
   module Http
     class Client
-      CYPHER_CLIENT_METHODS = %i[execute_cypher execute_batch_cypher].freeze
+      CYPHER_CLIENT_METHODS = %i[execute_cypher].freeze
       NODE_CLIENT_METHODS = %i[delete_node find_node_by find_nodes_by upsert_node].freeze
       RELATIONSHIP_CLIENT_METHODS = %i[delete_relationship upsert_relationship delete_relationship_on_primary_key].freeze
       CLIENT_METHODS = (CYPHER_CLIENT_METHODS + NODE_CLIENT_METHODS + RELATIONSHIP_CLIENT_METHODS).freeze
@@ -16,6 +16,11 @@ module Neo4j
           node_client = Http::NodeClient.new(cypher_client)
           relationship_client = Http::RelationshipClient.new(cypher_client)
           @default ||= new(cypher_client, node_client, relationship_client)
+        end
+
+        def in_batch &block
+          batch_client = Neo4j::Http::BatchClient
+          batch_client.execute_cypher yield(batch_client)
         end
       end